refactor: Extract sunshine accessor
